[Refactor#85] Banner(배너) 모델 클래스 banner 패키지로 이동
