+++ core/domain/src/main/java/com/ilsangtech/ilsang/core/domain/BannerRepository.kt
@@ -1,6 +1,6 @@
 package com.ilsangtech.ilsang.core.domain
 
-import com.ilsangtech.ilsang.core.model.Banner
+import com.ilsangtech.ilsang.core.model.banner.Banner
 
 interface BannerRepository {
     suspend fun getBanners(): List<Banner>
