+++ feature/home/src/main/java/com/ilsangtech/ilsang/feature/home/navigation/HomeNavigation.kt
@@ -3,7 +3,7 @@ package com.ilsangtech.ilsang.feature.home.navigation
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import com.ilsangtech.ilsang.core.model.Banner
+import com.ilsangtech.ilsang.core.model.banner.Banner
 import com.ilsangtech.ilsang.feature.home.HomeTabScreen
 import kotlinx.serialization.Serializable
